Fix `FuelMain` not being a decimal number

+++ EliteAPI.Status/Ship/ShipFuel.cs
@@ -5,7 +5,7 @@ namespace EliteAPI.Status.Ship;
 public readonly struct ShipFuel
 {
     [JsonProperty("FuelMain")]
-    public long FuelMain { get; init; }
+    public double FuelMain { get; init; }
 
     [JsonProperty("FuelReservoir")]
     public double FuelReservoir { get; init; }
